Moved freetype-render in a separate file
